+++ src/App.css
@@ -70,3 +70,7 @@ body::-webkit-scrollbar-thumb {
     border: solid 0.5px #787878;
     border-radius: 5px;
 }
+
+.img {
+    background: none;
+}
