+++ src/components/private-pages/Dashboard/PhotoText.js
