+++ src/components/public-pages/SignUp.js
@@ -4,12 +4,12 @@ import { Tooltip } from '@material-ui/core';
 import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
 import axios from 'axios';
 
-import validations from '../helperFunctions/validation';
-import { UserContext } from './Interface';
+import validations from '../../helperFunctions/validation';
+import { UserContext } from '../Interface';
 
 // Media Imports
-import { ReactComponent as Google } from '../images/google.svg';
-import { ReactComponent as Facebook } from '../images/facebook.svg';
+import { ReactComponent as Google } from '../../images/google.svg';
+import { ReactComponent as Facebook } from '../../images/facebook.svg';
 
 const SignUp = () => {
     const {_, dispatch} = useContext(UserContext);
